Input checkbox DOM events too weird
Just ran into every variation of trouble. To produce nice-looking
checkboxes, I was using markup where the <label> didn't wrap the
<input>. Then, to add another complication, Bootstrap's dropdown styles
require an <a>.

So, that meant I was attaching listeners in three different places.
Trying to deduplicate the events that were fired was proving way too
hard:

* React documentation says you can't prevent event propagation on the
  onChange handler, because it'll break the way they normalize browser
  events, and the UI state of the checkbox won't update.
* But I needed to prevent propagation in order to only fire a single
  event (given clicks on the wrapping <a> also need handling)
* React say a workaround to the UI not updating is to fire the change
  in a different context (with setTimeout). But you'll also need to
  call event.persist first, or stopping propagation won't do anything.

There's no mention of how to deal with wrapping elements with their own
listeners. Nor stuff like whether it's necessary to use a separate
handler for clicks on the label in order to preventDefault there (but
not on the input itself).

I'm sure I'm missing some finer point on how SyntheticEvents are
generated and used, but this just seems way too complicated. So, away
with the input.